Revert "PP-12853 Use axios or request retry for Apple Pay Merchant validation depending on configuration"
