+++ .secrets.baseline
@@ -114,7 +114,16 @@
         "filename": "app/controllers/web-payments/apple-pay/merchant-validation.controller.js",
         "hashed_secret": "1348b145fa1a555461c1b790a2f66614781091e9",
         "is_verified": false,
-        "line_number": 20
+        "line_number": 14
+      }
+    ],
+    "test/controllers/web-payments/apple-pay/merchant-validation.controller.test.js": [
+      {
+        "type": "Private Key",
+        "filename": "test/controllers/web-payments/apple-pay/merchant-validation.controller.test.js",
+        "hashed_secret": "1348b145fa1a555461c1b790a2f66614781091e9",
+        "is_verified": false,
+        "line_number": 68
       }
     ],
     "test/controllers/web-payments/apple-pay/normalise-apple-pay-payload.test.js": [
@@ -138,15 +147,6 @@
         "hashed_secret": "a269701da90143c626af966dfdaa775d848858b1",
         "is_verified": false,
         "line_number": 44
-      }
-    ],
-    "test/controllers/web-payments/apple-pay/old-merchant-validation.controller.test.js": [
-      {
-        "type": "Private Key",
-        "filename": "test/controllers/web-payments/apple-pay/old-merchant-validation.controller.test.js",
-        "hashed_secret": "1348b145fa1a555461c1b790a2f66614781091e9",
-        "is_verified": false,
-        "line_number": 69
       }
     ],
     "test/controllers/web-payments/google-pay/normalise-google-pay-payload.test.js": [
@@ -385,5 +385,5 @@
       }
     ]
   },
-  "generated_at": "2024-07-29T10:56:11Z"
+  "generated_at": "2024-07-22T12:54:11Z"
 }
