+++ test/controllers/web-payments/apple-pay/merchant-validation.controller.test.js
@@ -2,8 +2,6 @@
 
 const sinon = require('sinon')
 const proxyquire = require('proxyquire')
-const https = require('https')
-const { HttpsProxyAgent } = require('https-proxy-agent')
 
 const merchantDomain = 'www.pymnt.uk'
 const worldpayMerchantId = 'worldpay.merchant.id'
@@ -14,20 +12,19 @@ const stripeCertificate = 'A-STRIPE-CERTIFICATE'
 const stripeKey = 'A-STRIPE-KEY'
 const url = 'https://fakeapple.url'
 
-const appleResponse = { status: 200, data: { foo: 'bar' } }
+const appleResponse = { status: 200 }
+const appleResponseBody = { foo: 'bar' }
 
-function getControllerWithMocks (axiosMock) {
+function getControllerWithMocks (requestMock) {
   return proxyquire('../../../../app/controllers/web-payments/apple-pay/merchant-validation.controller', {
-    axios: axiosMock
+    requestretry: requestMock
   })
 }
 
 describe('Validate with Apple the merchant is legitimate', () => {
-  let res, sendSpy, axiosStub
+  let res, sendSpy
 
   beforeEach(() => {
-    delete process.env.HTTPS_PROXY
-
     process.env.APPLE_PAY_MERCHANT_DOMAIN = merchantDomain
     process.env.WORLDPAY_APPLE_PAY_MERCHANT_ID = worldpayMerchantId
     process.env.WORLDPAY_APPLE_PAY_MERCHANT_ID_CERTIFICATE = worldpayCertificate
@@ -35,205 +32,152 @@ describe('Validate with Apple the merchant is legitimate', () => {
     process.env.STRIPE_APPLE_PAY_MERCHANT_ID = stripeMerchantId
     process.env.STRIPE_APPLE_PAY_MERCHANT_ID_CERTIFICATE = stripeCertificate
     process.env.STRIPE_APPLE_PAY_MERCHANT_ID_CERTIFICATE_KEY = stripeKey
-    process.env.APPLE_PAY_MERCHANT_VALIDATION_VIA_AXIOS = 'true'
 
     sendSpy = sinon.spy()
     res = {
       status: sinon.spy(() => ({ send: sendSpy })),
       sendStatus: sinon.spy()
     }
-
-    axiosStub = sinon.stub().resolves(appleResponse)
   })
 
-  describe('when running locally with no proxy', () => {
-    it('should return a payload for a Worldpay payment if Merchant is valid', async () => {
-      const controller = getControllerWithMocks(axiosStub)
+  it('should return a payload for a Worldpay payment if Merchant is valid', async () => {
+    const mockRequest = sinon.stub().yields(null, appleResponse, appleResponseBody)
+    const controller = getControllerWithMocks(mockRequest)
 
-      const req = {
-        body: {
-          url,
-          paymentProvider: 'worldpay'
-        }
-      }
-      await controller(req, res)
-
-      sinon.assert.calledWith(axiosStub, sinon.match({
+    const req = {
+      body: {
         url,
-        method: 'post',
-        headers: { 'Content-Type': 'application/json' },
-        data: {
-          merchantIdentifier: worldpayMerchantId,
-          displayName: 'GOV.UK Pay',
-          initiative: 'web',
-          initiativeContext: merchantDomain
-        },
-        httpsAgent: sinon.match.instanceOf(https.Agent)
-      }))
+        paymentProvider: 'worldpay'
+      }
+    }
+    await controller(req, res)
 
-      const httpsAgentArg = axiosStub.getCall(0).args[0].httpsAgent
-      sinon.assert.match(httpsAgentArg.options, {
-        cert: sinon.match(cert => cert.includes(worldpayCertificate)),
-        key: sinon.match(key => key.includes(worldpayKey))
-      })
-
-      sinon.assert.calledWith(res.status, 200)
-      sinon.assert.calledWith(sendSpy, appleResponse.data)
+    sinon.assert.calledWith(mockRequest, {
+      url,
+      body: {
+        merchantIdentifier: worldpayMerchantId,
+        displayName: 'GOV.UK Pay',
+        initiative: 'web',
+        initiativeContext: merchantDomain
+      },
+      method: 'post',
+      json: true,
+      cert: `-----BEGIN CERTIFICATE-----
+${worldpayCertificate}
+-----END CERTIFICATE-----`,
+      key: `-----BEGIN PRIVATE KEY-----
+${worldpayKey}
+-----END PRIVATE KEY-----`
     })
+    sinon.assert.calledWith(res.status, 200)
+    sinon.assert.calledWith(sendSpy, appleResponseBody)
   })
 
-  describe('when there is a proxy', () => {
-    it('should return a payload for a Worldpay payment if Merchant is valid', async () => {
-      process.env.HTTPS_PROXY = 'https://fakeproxy.com'
-      const controller = getControllerWithMocks(axiosStub)
+  it('should return a payload for a Stripe payment if Merchant is valid', async () => {
+    const mockRequest = sinon.stub().yields(null, appleResponse, appleResponseBody)
+    const controller = getControllerWithMocks(mockRequest)
 
-      const req = {
-        body: {
-          url,
-          paymentProvider: 'worldpay'
-        }
-      }
-      await controller(req, res)
-
-      sinon.assert.calledWith(axiosStub, sinon.match({
+    const req = {
+      body: {
         url,
-        method: 'post',
-        headers: { 'Content-Type': 'application/json' },
-        data: {
-          merchantIdentifier: worldpayMerchantId,
-          displayName: 'GOV.UK Pay',
-          initiative: 'web',
-          initiativeContext: merchantDomain
-        },
-        httpsAgent: sinon.match.instanceOf(HttpsProxyAgent)
-      }))
-
-      const httpsAgentArg = axiosStub.getCall(0).args[0].httpsAgent
-      // sinon.assert.match(httpsAgentArg.options, { // not working locally for me
-      //   cert: sinon.match(cert => cert.includes(worldpayCertificate)),
-      //   key: sinon.match(key => key.includes(worldpayKey))
-      // })
-      sinon.match(httpsAgentArg.proxy, 'https://fakeproxy.com')
-      sinon.assert.calledWith(res.status, 200)
-      sinon.assert.calledWith(sendSpy, appleResponse.data)
-    })
-
-    it('should return a payload for a Stripe payment if Merchant is valid', async () => {
-      process.env.HTTPS_PROXY = 'https://fakeproxy.com'
-      const axiosStub = sinon.stub().resolves({ data: appleResponse.data, status: 200 })
-      const controller = getControllerWithMocks(axiosStub)
-
-      const req = {
-        body: {
-          url,
-          paymentProvider: 'stripe'
-        }
+        paymentProvider: 'stripe'
       }
-      await controller(req, res)
+    }
+    await controller(req, res)
 
-      sinon.assert.calledWith(axiosStub, sinon.match({
+    sinon.assert.calledWith(mockRequest, {
+      url,
+      body: {
+        merchantIdentifier: stripeMerchantId,
+        displayName: 'GOV.UK Pay',
+        initiative: 'web',
+        initiativeContext: merchantDomain
+      },
+      method: 'post',
+      json: true,
+      cert: `-----BEGIN CERTIFICATE-----
+${stripeCertificate}
+-----END CERTIFICATE-----`,
+      key: `-----BEGIN PRIVATE KEY-----
+${stripeKey}
+-----END PRIVATE KEY-----`
+    })
+    sinon.assert.calledWith(res.status, 200)
+    sinon.assert.calledWith(sendSpy, appleResponseBody)
+  })
+
+  it('should return 400 if no url is provided', async () => {
+    const mockRequest = sinon.stub().yields(null, appleResponse, appleResponseBody)
+    const controller = getControllerWithMocks(mockRequest)
+
+    const req = {
+      body: {
+        paymentProvider: 'worldpay'
+      }
+    }
+    await controller(req, res)
+    sinon.assert.calledWith(res.sendStatus, 400)
+  })
+
+  it('should return a payload for a Sandbox payment if Merchant is valid', async () => {
+    const mockRequest = sinon.stub().yields(null, appleResponse, appleResponseBody)
+    const controller = getControllerWithMocks(mockRequest)
+
+    const req = {
+      body: {
         url,
-        method: 'post',
-        headers: { 'Content-Type': 'application/json' },
-        data: {
-          merchantIdentifier: stripeMerchantId,
-          displayName: 'GOV.UK Pay',
-          initiative: 'web',
-          initiativeContext: merchantDomain
-        },
-        httpsAgent: sinon.match.instanceOf(HttpsProxyAgent)
-      }))
-
-      const httpsAgentArg = axiosStub.getCall(0).args[0].httpsAgent
-      sinon.match(httpsAgentArg.proxy, 'https://fakeproxy.com')
-      // sinon.assert.match(httpsAgentArg.connectOpts, { // not working locally for me
-      //   cert: sinon.match(cert => cert.includes(stripeCertificate)),
-      //   key: sinon.match(key => key.includes(stripeKey))
-      // })
-
-      sinon.assert.calledWith(res.status, 200)
-      sinon.assert.calledWith(sendSpy, appleResponse.data)
-    })
-
-    it('should return 400 if no url is provided', async () => {
-      const axiosStub = sinon.stub().resolves({ data: appleResponse.data, status: 200 })
-      const controller = getControllerWithMocks(axiosStub)
-
-      const req = {
-        body: {
-          paymentProvider: 'worldpay'
-        }
+        paymentProvider: 'sandbox'
       }
-      await controller(req, res)
-      sinon.assert.calledWith(res.sendStatus, 400)
-      sinon.assert.notCalled(axiosStub)
+    }
+    await controller(req, res)
+
+    sinon.assert.calledWith(mockRequest, {
+      url,
+      body: {
+        merchantIdentifier: worldpayMerchantId,
+        displayName: 'GOV.UK Pay',
+        initiative: 'web',
+        initiativeContext: merchantDomain
+      },
+      method: 'post',
+      json: true,
+      cert: `-----BEGIN CERTIFICATE-----
+${worldpayCertificate}
+-----END CERTIFICATE-----`,
+      key: `-----BEGIN PRIVATE KEY-----
+${worldpayKey}
+-----END PRIVATE KEY-----`
     })
+    sinon.assert.calledWith(res.status, 200)
+    sinon.assert.calledWith(sendSpy, appleResponseBody)
+  })
 
-    it('should return a payload for a Sandbox payment if Merchant is valid', async () => {
-      process.env.HTTPS_PROXY = 'https://fakeproxy.com'
-      const axiosStub = sinon.stub().resolves({ data: appleResponse.data, status: 200 })
-      const controller = getControllerWithMocks(axiosStub)
+  it('should return 400 for unexpected payment provider', async () => {
+    const mockRequest = sinon.stub().yields(null, appleResponse, appleResponseBody)
+    const controller = getControllerWithMocks(mockRequest)
 
-      const req = {
-        body: {
-          url,
-          paymentProvider: 'sandbox'
-        }
-      }
-      await controller(req, res)
-
-      sinon.assert.calledWith(axiosStub, sinon.match({
+    const req = {
+      body: {
         url,
-        method: 'post',
-        headers: { 'Content-Type': 'application/json' },
-        data: {
-          merchantIdentifier: worldpayMerchantId,
-          displayName: 'GOV.UK Pay',
-          initiative: 'web',
-          initiativeContext: merchantDomain
-        },
-        httpsAgent: sinon.match.instanceOf(HttpsProxyAgent)
-      }))
-
-      const httpsAgentArg = axiosStub.getCall(0).args[0].httpsAgent
-      sinon.match(httpsAgentArg.proxy, 'https://fakeproxy.com')
-      // sinon.assert.match(httpsAgentArg.connectOpts, { // not working locally for me
-      //   cert: sinon.match(cert => cert.includes(worldpayCertificate)),
-      //   key: sinon.match(key => key.includes(worldpayKey))
-      // })
-
-      sinon.assert.calledWith(res.status, 200)
-      sinon.assert.calledWith(sendSpy, appleResponse.data)
-    })
-
-    it('should return 400 for unexpected payment provider', async () => {
-      const axiosStub = sinon.stub().resolves({ data: appleResponse.data, status: 200 })
-      const controller = getControllerWithMocks(axiosStub)
-
-      const req = {
-        body: {
-          url,
-          paymentProvider: 'mystery'
-        }
+        paymentProvider: 'mystery'
       }
-      await controller(req, res)
-      sinon.assert.calledWith(res.sendStatus, 400)
-      sinon.assert.notCalled(axiosStub)
-    })
+    }
+    await controller(req, res)
+    sinon.assert.calledWith(res.sendStatus, 400)
+  })
 
-    it('should return an error if Apple Pay returns an error', async () => {
-      const axiosErrorStub = sinon.stub().rejects(new Error('Whatever error from Apple Pay'))
-      const controller = getControllerWithMocks(axiosErrorStub)
+  it('should return an error if Apple Pay returns an error', async () => {
+    const mockRequest = sinon.stub().yields(new Error(), appleResponse, appleResponseBody)
+    const controller = getControllerWithMocks(mockRequest)
 
-      const req = {
-        body: {
-          url,
-          paymentProvider: 'worldpay'
-        }
+    const req = {
+      body: {
+        url,
+        paymentProvider: 'worldpay'
       }
-      await controller(req, res)
-      sinon.assert.calledWith(res.status, 500)
-      sinon.assert.calledWith(sendSpy, 'Apple Pay Error')
-    })
+    }
+    await controller(req, res)
+    sinon.assert.calledWith(res.status, 500)
+    sinon.assert.calledWith(sendSpy, appleResponseBody)
   })
 })
